fixed jpa and jackson_error

+++ src/main/java/com/mike/courseregistrationapp/Repo/CourseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseRepository extends JpaRepository<CourseDetails, Integer> {
+public interface CourseRepository extends JpaRepository<CourseDetails, Long> {
 }
