+++ src/main/java/com/mike/courseregistrationapp/Repo/OfferingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfferingRepository extends JpaRepository<Offering, Integer> {
+public interface OfferingRepository extends JpaRepository<Offering, Long> {
 }
